requirements.txt
# --------- pytorch --------- #
torch==2.2.2
torchvision>=0.15.0
lightning==2.2
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
black
fire
overrides
gdown
# sh            # for running bash commands in some tests (linux/macos only)

# --------- Warp ConvNet -------- #
# git+ssh://git@gitlab-master.nvidia.com:12051/3dmmllm/warp.git
jaxtyping
--find-links https://data.pyg.org/whl/torch-2.2.0+cu121.html
torch-scatter

# --------- RegionPLC --------- #
numpy
spconv-cu120
easydict
pyyaml
tqdm
scipy
plyfile
imageio
matplotlib
transformers
clip @ git+https://github.com/openai/CLIP.git
einops
open3d==0.18.0
natsort
scikit-image
open_clip_torch

# --------- Segment3D --------- #
--extra-index-url https://pypi.nvidia.com
cuml-cu12   # for DBSCAN
loguru
albumentations
pyviz3d
